feat(ui): add WIP QMenu support and refactor style architecture
Implement custom painting for QMenu via a new `MenuDrawer` and
restructure the styling system to resolve circular dependencies.

Key changes:
- Added `MenuDrawer` to handle custom `QPainter` rendering for `QMenu`
  elements (panels, frames, and menu items).
- Introduced `style_types.py` to host `StyleDict`, `StyleData`, and
  singleton accessors, decoupling them from the main `style.py` to
  eliminate circular imports between components and the style engine.
- Moved `ComboBoxItemDelegate`, `TableItemDelegate`, and
  `TreeViewItemDelegate` from the `drawers` module to their respective
  component files to improve modularity.
- Added `QMenu` configuration to `ayon_style.json`.
- Added visual regression tests for the new menu styling.
