+++ client/ayon_ui_qt/ayon_style.json
@@ -546,6 +546,33 @@
                 }
             }
         },
+        "QMenu": {
+            "states": ["hover", "disabled"],
+            "default-variant": "default",
+            "variants": {
+                "default": {
+                    "background-color": "--md-sys-color-surface-container-low-dark",
+                    "color": "--md-sys-color-on-surface-dark",
+                    "border-radius": 2,
+                    "border-width": 1,
+                    "border-color": "--md-sys-color-outline-variant-dark",
+                    "panel-padding": [4, 4],
+                    "item-padding": [6, 12],
+                    "item-radius": 4,
+                    "item-spacing": 4,
+                    "icon-size": 16,
+                    "separator-color": "--md-sys-color-outline-variant-dark",
+                    "separator-height": 1,
+                    "hover": {
+                        "background-color": "--md-sys-color-surface-container-highest-dark",
+                        "color": "--md-sys-color-on-surface-dark"
+                    },
+                    "disabled": {
+                        "opacity": 0.5
+                    }
+                }
+            }
+        },
         "QLabel": {
             "states": ["disabled", "hover"],
             "default-variant": "default",
