+++ client/ayon_ui_qt/components/slider.py
@@ -6,7 +6,8 @@ from qtpy.QtCore import QRectF, Qt, Signal
 from qtpy.QtGui import QBrush, QColor, QPainter, QPaintEvent, QPen, QPixmap
 from qtpy.QtWidgets import QSlider, QWidget
 
-from ..style import StyleDict, get_ayon_style
+from ..style import get_ayon_style
+from ..style_types import StyleDict
 from ..variants import AYSliderVariants
 from .container import AYContainer
 from .label import AYLabel
