+++ client/ayon_ui_qt/components/tree_view.py
@@ -2,18 +2,38 @@
 
 from __future__ import annotations
 
-from qtpy.QtCore import QEvent, QItemSelection, Qt, Signal
+
+from qtpy.QtCore import (
+    QEvent,
+    QItemSelection,
+    QModelIndex,
+    QPersistentModelIndex,
+    QRect,
+    QSize,
+    Qt,
+    Signal,
+)
 from qtpy.QtGui import (
+    QBrush,
     QColor,
     QCursor,
+    QIcon,
     QMouseEvent,
     QPainter,
     QPaintEvent,
     QPalette,
 )
-from qtpy.QtWidgets import QStyle, QStyleOption, QTreeView, QWidget
+from qtpy.QtWidgets import (
+    QStyle,
+    QStyledItemDelegate,
+    QStyleOption,
+    QStyleOptionViewItem,
+    QTreeView,
+    QWidget,
+)
 
-from ..style import TreeViewItemDelegate, enum_to_str, get_ayon_style
+from ..style import enum_to_str, get_ayon_style
+from ..style_types import StyleData
 from ..variants import QTreeViewVariants
 from .scroll_area import AYScrollBar
 from .style_mixin import StyleMixin
@@ -206,6 +226,198 @@ class AYTreeView(StyleMixin, QTreeView):
         """
         super().selectionChanged(selected, deselected)
         self.selection_changed.emit(selected, deselected)
+
+
+class TreeViewItemDelegate(StyleMixin, QStyledItemDelegate):
+    """Item delegate for AYTreeView that paints directly, bypassing QSS.
+
+    Reads style data from the QTreeView style entry to draw item
+    backgrounds (hover, selected) and text/icons.  The paint() method
+    uses raw QPainter calls so that a parent-level QStyleSheet cannot
+    intercept and override the colours.
+
+    Args:
+        parent: The parent widget (expected to be an AYTreeView instance).
+        style_model: StyleData instance providing colour/dimension data.
+        variant: The variant string used to look up the correct style.
+    """
+
+    def __init__(
+        self,
+        parent: QWidget | None = None,
+        style_model: StyleData | None = None,
+        variant: str = "default",
+    ) -> None:
+        super().__init__(parent)
+        self._style_model = style_model
+        self._variant_str = variant
+        self._icon_cache: dict[str, QIcon] = {}
+
+    def _tv_styles(self) -> dict[str, dict]:
+        """Return *base*, *hover* and *selected* style dicts at once."""
+        if self._style_model is None:
+            return {"base": {}, "hover": {}, "selected": {}}
+        return self._style_model.get_styles(
+            "QTreeView",
+            self._variant_str,
+            ["base", "hover", "selected"],
+        )
+
+    def initStyleOption(
+        self,
+        option: QStyleOptionViewItem,
+        index: QModelIndex | QPersistentModelIndex,
+    ) -> None:
+        """Initialize the style option with the default implementation, then
+        override any properties needed for our custom painting.
+
+        Args:
+            option: The style option to initialize.
+            index: The model index of the item.
+        """
+        super().initStyleOption(option, index)
+        option.font = self.font()
+        option.fontMetrics = self.fontMetrics()
+
+    def sizeHint(
+        self,
+        option: QStyleOptionViewItem,
+        index: QModelIndex | QPersistentModelIndex,
+    ) -> QSize:
+        """Return a fixed row height from the style data.
+
+        Args:
+            option: The style option for the item.
+            index: The model index of the item.
+
+        Returns:
+            The size hint for the item.
+        """
+        if self._style_model:
+            style = self._style_model.get_style("QTreeView", self._variant_str)
+            h = int(style.get("item-height", 28))
+        else:
+            h = 28
+        return QSize(option.rect.width(), h)
+
+    def paint(
+        self,
+        painter: QPainter,
+        option: QStyleOptionViewItem,
+        index: QModelIndex | QPersistentModelIndex,
+    ) -> None:
+        """Paint a tree-view item directly, bypassing QStyle completely.
+
+        Args:
+            painter: The QPainter to use.
+            option: The style option for the item.
+            index: The model index of the item.
+        """
+        painter.save()
+        painter.setRenderHint(QPainter.RenderHint.Antialiasing)
+
+        opt = QStyleOptionViewItem(option)
+        self.initStyleOption(opt, index)
+
+        state = opt.state
+        is_selected = bool(state & QStyle.StateFlag.State_Selected)
+        is_hovered = bool(state & QStyle.StateFlag.State_MouseOver)
+
+        styles = self._tv_styles()
+        base_style = styles["base"]
+        hover_style = styles["hover"]
+        selected_style = styles["selected"]
+
+        item_padding = base_style.get("item-padding", [4, 8])
+        icon_text_spacing = int(base_style.get("icon-text-spacing", 6))
+
+        # --- background ------------------------------------------------
+        if is_selected:
+            bg_color = QColor(
+                selected_style.get(
+                    "background-color",
+                    base_style.get("background-color", "transparent"),
+                )
+            )
+        elif is_hovered:
+            bg_color = QColor(
+                hover_style.get(
+                    "background-color",
+                    base_style.get("background-color", "transparent"),
+                )
+            )
+        else:
+            bg_color = QColor(
+                base_style.get("background-color", "transparent")
+            )
+
+        painter.setBrush(QBrush(bg_color))
+        painter.setPen(Qt.PenStyle.NoPen)
+        painter.drawRect(opt.rect)
+
+        # --- text colour -----------------------------------------------
+        if is_selected:
+            text_color = QColor(
+                selected_style.get(
+                    "color",
+                    base_style.get("color", "#f4f5f5"),
+                )
+            )
+        else:
+            text_color = QColor(base_style.get("color", "#f4f5f5"))
+
+        # disabled dimming
+        if not (state & QStyle.StateFlag.State_Enabled):
+            text_color.setAlpha(
+                int(
+                    text_color.alpha()
+                    * base_style.get("disabled-opacity", 0.5)
+                )
+            )
+
+        # --- icon + text layout ----------------------------------------
+        content_rect = QRect(opt.rect).adjusted(
+            item_padding[1],
+            item_padding[0],
+            -item_padding[1],
+            -item_padding[0],
+        )
+        content_left = content_rect.left()
+
+        if not opt.icon.isNull():
+            icon_size = opt.decorationSize
+            icon_rect = QRect(
+                content_left,
+                opt.rect.center().y() - icon_size.height() // 2,
+                icon_size.width(),
+                icon_size.height(),
+            )
+            mode = (
+                QIcon.Mode.Normal
+                if state & QStyle.StateFlag.State_Enabled
+                else QIcon.Mode.Disabled
+            )
+            opt.icon.paint(
+                painter,
+                icon_rect,
+                Qt.AlignmentFlag.AlignCenter,
+                mode,
+            )
+            content_left = icon_rect.right() + icon_text_spacing
+
+        if opt.text:
+            text_rect = QRect(opt.rect)
+            text_rect.setLeft(content_left)
+            text_rect.setRight(content_rect.right())
+            painter.setPen(text_color)
+            painter.setFont(opt.font)
+            painter.drawText(
+                text_rect,
+                Qt.AlignmentFlag.AlignVCenter | Qt.AlignmentFlag.AlignLeft,
+                opt.text,
+            )
+
+        painter.restore()
 
 
 # =============================================================================
